renaming and exporting rpc and signature provider
