+++ src/eosjs-jsonrpc.ts
@@ -3,11 +3,10 @@
  */
 
 // copyright defined in eosjs/LICENSE.txt
-import { base64ToBinary } from "./eosjs-numeric";
-import { convertLegacyPublicKeys } from "./eosjs-numeric";
+import { base64ToBinary, convertLegacyPublicKeys } from "./eosjs-numeric";
 import { RpcError } from "./eosjs-rpcerror";
 import { AbiProvider, AuthorityProvider, AuthorityProviderArgs, BinaryAbi } from "./api-interfaces";
-import { GetAbiResult, GetBlockResult, GetCodeResult, GetInfoResult, GetRawCodeAndAbiResult, PushTransactionArgs } from "./jsonrpc-interfaces";
+import { GetAbiResult, GetBlockResult, GetCodeResult, GetInfoResult, GetRawCodeAndAbiResult, PushTransactionArgs } from "./rpc-interfaces";
 
 function arrayToHex(data: Uint8Array) {
     let result = "";
