+++ src/index.ts
@@ -1,9 +1,9 @@
 import "babel-polyfill";
 import { Api } from "./eosjs-api";
 import * as ApiInterfaces from "./api-interfaces"
-import * as Rpc from "./eosjs-jsonrpc";
-import * as RpcInterfaces from "./jsonrpc-interfaces"
-import SignatureProvider from "./eosjs-jssig";
+import { JsonRpc } from "./eosjs-jsonrpc";
+import * as RpcInterfaces from "./rpc-interfaces"
+import { JsSignatureProvider } from "./eosjs-jssig";
 import * as Serialize from "./eosjs-serialize";
 
-export { Api, ApiInterfaces, SignatureProvider, Rpc, RpcInterfaces, Serialize };
+export { Api, ApiInterfaces, JsonRpc, RpcInterfaces, JsSignatureProvider, Serialize };
